+++ src/components/Cell/Cell.tsx
@@ -1,14 +1,15 @@
 import * as React from 'react';
 import { classNames } from '../../lib/classNames';
 import { getClassName } from '../../helpers/getClassName';
-import { Touch, TouchEvent } from '../Touch/Touch';
 import { ANDROID, IOS, VKCOM } from '../../lib/platform';
-import { Icon24Reorder, Icon24ReorderIos, Icon24CheckCircleOn, Icon24CheckCircleOff, Icon24CheckBoxOff, Icon24CheckBoxOn } from '@vkontakte/icons';
+import { Icon24CheckCircleOn, Icon24CheckCircleOff, Icon24CheckBoxOff, Icon24CheckBoxOn } from '@vkontakte/icons';
 import SimpleCell, { SimpleCellProps } from '../SimpleCell/SimpleCell';
 import { HasPlatform } from '../../types';
 import { Removable, RemovableProps } from '../Removable/Removable';
 import { usePlatform } from '../../hooks/usePlatform';
+import { useDraggable } from './useDraggable';
 import { ListContext } from '../List/ListContext';
+import { CellDragger } from '../CellDragger/CellDragger';
 import './Cell.css';
 
 export interface CellProps extends SimpleCellProps, HasPlatform, RemovableProps {
@@ -44,6 +45,10 @@ export interface CellProps extends SimpleCellProps, HasPlatform, RemovableProps 
    * есть рабочий пример с обработкой этих чисел и перерисовкой списка.
    */
   onDragFinish?: ({ from, to }: { from: number; to: number }) => void;
+  /**
+   * aria-label для кнопки перетаскивания ячейки
+   */
+  draggerLabel?: string;
 }
 
 export const Cell: React.FC<CellProps> = ({
@@ -65,6 +70,7 @@ export const Cell: React.FC<CellProps> = ({
   checked,
   defaultChecked,
   getRootRef,
+  draggerLabel = 'Перенести ячейку',
   ...restProps
 }: CellProps) => {
   // TODO: удалить эту и следующие 7 строк перед 5.0.0
@@ -79,94 +85,9 @@ export const Cell: React.FC<CellProps> = ({
   const selectable = mode === 'selectable';
   const removable = mode === 'removable';
 
-  const rootElRef = React.useRef(null);
   const platform = usePlatform();
 
-  const [dragging, setDragging] = React.useState<boolean>(false);
-
-  const [siblings, setSiblings] = React.useState<HTMLElement[]>(undefined);
-  const [dragStartIndex, setDragStartIndex] = React.useState<number>(undefined);
-  const [dragEndIndex, setDragEndIndex] = React.useState<number>(undefined);
-  const [dragShift, setDragShift] = React.useState<number>(0);
-  const [dragDirection, setDragDirection] = React.useState<'down' | 'up'>(undefined);
-
-  const onDragStart = () => {
-    const rootEl = rootElRef?.current;
-
-    setDragging(true);
-
-    const _siblings: HTMLElement[] = Array.from(rootEl.parentElement.childNodes);
-    const rootElIdx = _siblings.indexOf(rootEl);
-
-    setDragStartIndex(rootElIdx);
-    setDragEndIndex(rootElIdx);
-    setSiblings(_siblings);
-    setDragShift(0);
-  };
-
-  const onDragMove = (e: TouchEvent) => {
-    e.originalEvent.preventDefault();
-
-    const rootEl = rootElRef?.current;
-
-    rootEl.style.transform = `translateY(${e.shiftY}px)`;
-    setDragDirection(dragShift - e.shiftY < 0 ? 'down' : 'up');
-    setDragShift(e.shiftY);
-    setDragEndIndex(dragStartIndex);
-
-    siblings.forEach((sibling: HTMLElement, siblingIndex: number) => {
-      const rootGesture = rootEl.getBoundingClientRect();
-
-      const siblingGesture = sibling.getBoundingClientRect();
-
-      if (dragStartIndex < siblingIndex) {
-        if (rootGesture.bottom > siblingGesture.top + siblingGesture.height / 2) {
-          if (dragDirection === 'down') {
-            sibling.style.transform = 'translateY(-100%)';
-          }
-
-          setDragEndIndex((dragEndIndex) => dragEndIndex + 1);
-        }
-        if (rootGesture.top < siblingGesture.bottom - siblingGesture.height / 2 && dragDirection === 'up') {
-          sibling.style.transform = 'translateY(0)';
-        }
-      } else if (dragStartIndex > siblingIndex) {
-        if (rootGesture.top < siblingGesture.bottom - siblingGesture.height / 2) {
-          if (dragDirection === 'up') {
-            sibling.style.transform = 'translateY(100%)';
-          }
-
-          setDragEndIndex((dragEndIndex) => dragEndIndex - 1);
-        }
-        if (rootGesture.bottom > siblingGesture.top + siblingGesture.height / 2 && dragDirection === 'down') {
-          sibling.style.transform = 'translateY(0)';
-        }
-      }
-    });
-  };
-
-  const onDragEnd = () => {
-    const [from, to] = [dragStartIndex, dragEndIndex];
-
-    siblings.forEach((sibling: HTMLElement) => {
-      sibling.style.transform = null;
-    });
-
-    setSiblings(undefined);
-    setDragEndIndex(undefined);
-    setDragStartIndex(undefined);
-    setDragDirection(undefined);
-    setDragShift(undefined);
-
-    setDragging(false);
-
-    onDragFinish && onDragFinish({ from, to });
-  };
-
-  const onDragClick = (e: React.MouseEvent) => {
-    e.nativeEvent.stopPropagation();
-    e.preventDefault();
-  };
+  const { dragging, rootElRef, ...draggableProps } = useDraggable({ onDragFinish });
 
   const { toggleDrag } = React.useContext(ListContext);
   React.useEffect(() => {
@@ -180,6 +101,11 @@ export const Cell: React.FC<CellProps> = ({
   const IconOff = platform === ANDROID ? Icon24CheckBoxOff : Icon24CheckCircleOff;
   const IconOn = platform === ANDROID ? Icon24CheckBoxOn : Icon24CheckCircleOn;
 
+  let dragger;
+  if (draggable) {
+    dragger = <CellDragger aria-label={draggerLabel} {...draggableProps} />;
+  }
+
   const simpleCell = (
     <SimpleCell
       {...restProps}
@@ -188,15 +114,7 @@ export const Cell: React.FC<CellProps> = ({
       htmlFor={selectable ? name : undefined}
       before={
         <React.Fragment>
-          {(platform === ANDROID || platform === VKCOM) && draggable && (
-            <Touch
-              vkuiClass="Cell__dragger"
-              onStart={onDragStart}
-              onMoveY={onDragMove}
-              onEnd={onDragEnd}
-              onClick={onDragClick}
-            ><Icon24Reorder /></Touch>
-          )}
+          {draggable && (platform === ANDROID || platform === VKCOM) && dragger}
           {selectable && (
             <React.Fragment>
               <input
@@ -219,15 +137,7 @@ export const Cell: React.FC<CellProps> = ({
       }
       after={
         <React.Fragment>
-          {platform === IOS && draggable && (
-            <Touch
-              vkuiClass="Cell__dragger"
-              onStart={onDragStart}
-              onMoveY={onDragMove}
-              onEnd={onDragEnd}
-              onClick={onDragClick}
-            ><Icon24ReorderIos /></Touch>
-          )}
+          {draggable && platform === IOS && dragger}
           {after}
         </React.Fragment>
       }
